requirements.txt

fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
opencv-python==4.10.0.84
ffmpeg-python==0.2.0
numpy==2.2.1

pillow==11.0.0


torch==2.5.1
torchvision==0.20.1

ultralytics==8.3.53



